refactor(tests): split discord test_bot into focused suites
The 370 LOC test_bot.py covered four collaborators behind one filename
— DiscordSlashRegistrar, DiscordAgentRouter, the registry alias
behaviour, and four discord-library smoke tests. Split:

- test_slash_register.py — TestNormalizeName, TestBuildSignature,
  TestRegisterAll. fake_command + _patch_registry helpers replace
  the per-test boilerplate.
- test_agent_router.py — TestHandleDm. _stub_executor / _stub_strategy
  / _stub_no_strategy helpers wrap the mock setup.
- test_command_registry.py picks up the multi-word alias regression
  test (rule 34 / rule34).

The four "library smoke" tests (discord.File construction,
discord.Embed.set_image, str slicing) and the VideoContent isinstance
checks add no coverage of our code, so they're dropped per the testing
rule against smoke tests. test_bot.py is removed since DiscordBot's
behaviour is fully exercised through the registrar and router suites.
